address outstanding code review comments
