+++ nuxt_src/layouts/default.vue
@@ -7,7 +7,6 @@
     </lazy-component>
     <flow-sns v-if="$mq === 'lg'" />
     <the-footer />
-    <script id="facebook-jssdk" type="text/javascript" :src="`https://connect.facebook.net/${$t('locale_iso')}/sdk.js#xfbml=1&version=v3.0`" />
   </div>
 </template>
